requirements-api.txt
# Lightweight API Service Dependencies
fastapi==0.117.1
uvicorn[standard]==0.37.0
pydantic==2.11.9
pydantic-settings==2.11.0
python-dotenv==1.1.1
chromadb==1.1.0
sqlite-utils==3.38
# New Vertex AI SDK
google-genai
google-cloud-aiplatform
httpx==0.28.1
numpy
qdrant-client==1.9.2
